jwt_auth_project/requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.7.2
Django                        4.2.3
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
Pillow                        10.0.0
pip                           23.1.2
PyJWT                         2.8.0
pytz                          2023.3
setuptools                    68.0.0
sqlparse                      0.4.4
tzdata                        2023.3
wheel                         0.40.0
